pyproject.toml
[tool.poetry]
name = "arbitrage-bot"
version = "0.1.0"
description = "A cryptocurrency arbitrage bot that monitors multiple exchanges for price differences"
authors = ["Matus Klecka <klecka.matus@gmail.com>"]
readme = "README.md"
packages = [{include = "src"}]

[tool.poetry.dependencies]
python = "^3.9"
ccxt = "^4.1.13"
pandas = "^2.1.4"
numpy = "^1.26.2"
requests = "^2.31.0"
python-dotenv = "^1.0.0"
pytz = "^2023.3"
aiohttp = "^3.9.1"
asyncio = "^3.4.3"
websockets = "^12.0"
pydantic = "^2.5.2"
loguru = "^0.7.2"

[tool.poetry.group.dev.dependencies]
pytest = "^7.4.3"
pytest-asyncio = "^0.21.1"
pytest-cov = "^4.1.0"
black = "^23.11.0"
isort = "^5.12.0"
mypy = "^1.7.1"
flake8 = "^6.1.0"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

[tool.black]
line-length = 100
target-version = ['py39']
include = '\.pyi?$'

[tool.isort]
profile = "black"
multi_line_output = 3
line_length = 100

[tool.mypy]
python_version = "3.9"
warn_return_any = true
warn_unused_configs = true
disallow_untyped_defs = true
disallow_incomplete_defs = true

[tool.pytest.ini_options]
testpaths = ["tests"]
python_files = ["test_*.py"]
asyncio_mode = "auto" 